+++ Common/Core/UnitTests/ByteArrayTest.cpp
@@ -1,9 +1,9 @@
 
 #include "gtest/gtest.h"
-#include "Utilities/ByteArray.h"
+#include "Core/ByteArray.h"
 
 using namespace Realisim;
-using namespace Utilities;
+using namespace Core;
 
 const std::string referenceString = "the quick brown fox jumps over the lazy dog.";
 const std::string loremIpsum = "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.";
